FactCheckerProject/FactCheckerFunction/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
openai
requests
azure-ai-projects==1.0.0
azure.ai.agents
azure.identity